fix all the issue

+++ src/utils/cloudinary.js
@@ -2,9 +2,9 @@ import {v2 as cloudinary} from "cloudinary"
 import fs from 'fs'
 //import { Upload } from "../middleware/multer";
 cloudinary.config({ 
-  cloud_name:"dvm0kbwo4", 
-  api_key:process.env.CLOUDINARY_API_KEY, 
-  api_secret:process.env.CLOUDINARY_API_SECRET
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+  api_key: process.env.CLOUDINARY_API_KEY, 
+  api_secret: process.env.CLOUDINARY_API_SECRET
 });
 
 
@@ -27,11 +27,11 @@ const uploadOnCloudinary = async (localFilePath) => {
         console.log(response,"---cloudinary response")
         // file has been uploaded successfull
         //console.log("file is uploaded on cloudinary ", response.url);
-        fs.unlinkSync(localFilePath)
+        if (fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath)
         return response;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+        if (fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath)
         return null;
     }
 }
